backend/requirements.txt
paramiko
langchain
langgraph
langchain-openai
fastapi
uvicorn
python-dotenv